+++ src/mock/v1.2/channelService-1_2.ts
@@ -1,7 +1,7 @@
 import { Channel, DesktopAgent } from "fdc3_1_2";
 import constants from "../../constants";
 import { ChannelsAppConfig } from "../../test/common/channel-control";
-import { AppControlContext } from "../../test/common/common-types";
+import { AppControlContext } from "../../common-types";
 import { commands, channelType } from "../constants";
 declare let fdc3: DesktopAgent;
