+++ src/mock/v2.0/channelService-2_0.ts
@@ -1,7 +1,7 @@
 import { Channel, DesktopAgent } from "fdc3_2_0";
 import constants from "../../constants";
 import { ChannelsAppConfig } from "../../test/common/channel-control";
-import { AppControlContext } from "../../test/common/common-types";
+import { AppControlContext } from "../../common-types";
 import { commands, channelType } from "../constants";
 declare let fdc3: DesktopAgent;
