+++ src/mock/v2.0/intent-b.ts
@@ -2,7 +2,7 @@ import { closeWindowOnCompletion, onFdc3Ready } from "./mock-functions";
 import { DesktopAgent } from "fdc3_2_0/dist/api/DesktopAgent";
 import { sendContextToTests } from "../v2.0/mock-functions";
 import { wait } from "../../utils";
-import { AppControlContext, IntentUtilityContext } from "../../test/common/common-types";
+import { AppControlContext, IntentUtilityContext } from "../../common-types";
 declare let fdc3: DesktopAgent;
 onFdc3Ready().then(async () => {
   await closeWindowOnCompletion();
