+++ src/mock/v2.0/open-a.ts
@@ -1,7 +1,7 @@
 import { closeWindowOnCompletion, onFdc3Ready } from "./mock-functions";
 import { DesktopAgent } from "fdc3_2_0/dist/api/DesktopAgent";
 import { sendContextToTests } from "../v2.0/mock-functions";
-import { AppControlContext } from "../../test/common/common-types";
+import { AppControlContext } from "../../common-types";
 
 declare let fdc3: DesktopAgent;
 onFdc3Ready().then(async () => {
