+++ src/test/v2.0/advanced/channels-support-2.0.ts
@@ -3,7 +3,7 @@ import { Channel, Context, Listener, DesktopAgent } from "fdc3_2_0";
 import constants from "../../../constants";
 import { wait } from "../../../utils";
 import { ChannelControl, ChannelsAppConfig, ChannelsAppContext } from "../../common/channel-control";
-import { AppControlContext } from "../../common/common-types";
+import { AppControlContext } from "../../../common-types";
 
 declare let fdc3: DesktopAgent;
