+++ src/test/v2.0/advanced/intent-support-2.0.ts
@@ -4,7 +4,7 @@ import { Context, DesktopAgent, getOrCreateChannel } from "fdc3_2_0";
 import { APIDocumentation2_0 } from "../apiDocuments-2.0";
 import constants from "../../../constants";
 import { sleep, wait, wrapPromise } from "../../../utils";
-import { AppControlContext, IntentUtilityContext } from "../../common/common-types";
+import { AppControlContext, IntentUtilityContext } from "../../../common-types";
 
 declare let fdc3: DesktopAgent;
 const raiseIntentDocs = "\r\nDocumentation: " + APIDocumentation2_0.raiseIntent + "\r\nCause";
